fix: change data type

+++ functions_nested_loops/6-abs.c
@@ -11,7 +11,7 @@
  */
 int _abs(int n)
 {
-	short abs;
+	int abs;
 
 	if (n < 0)
 	{
